pull highlight styles search into helper function, run lint
